First Draft of Cog setup.
Complete restructering of code, addition of multiple helper functions, code streamlining.
