+++ cogs/helper_cogs.py
@@ -1,4 +1,6 @@
-from discord.ext import commands
+# this cog provides the utiltiy functions used by the commands to actually make changes
+# to avoid issues with multiple connections to the DB, anything that uses the mysql DB will exist in this cog even if it is only used once elsewhere
+
 #import statements
 import os
 import discord
@@ -16,6 +18,7 @@ import numpy as np
 
 class helper_cogs(commands.Cog, name='Utilities'): 
     
+    # this method is called on loading of the cog.
     def __init__(self, bot):
         self.bot = bot
 
@@ -23,7 +26,6 @@ class helper_cogs(commands.Cog, name='Utilities'):
         global mydb
         global sun_chan_code
         global raid_chan_code
-        global admin_role_code
         global bot_admin_code
 
         #create DB connection
@@ -42,7 +44,6 @@ class helper_cogs(commands.Cog, name='Utilities'):
         sun_chan_code = int(os.getenv('SUN_CHAN_CODE'))
         #raid_chan_code = int(os.getenv('RAID_CHAN_CODE')) 
         raid_chan_code = int(os.getenv('TEST_RAID_CHAN'))  #secondary channel for testing
-        admin_role_code = int(os.getenv('ADMIN_ROLE_CODE'))
         bot_admin_code = int(os.getenv('BOT_ADMIN_CODE'))
 
 
@@ -85,7 +86,6 @@ class helper_cogs(commands.Cog, name='Utilities'):
         await raid_message.edit(content = f'{details}{primaries}{backups}{requirements}')
 
     #helper function to ask user what raid they want to run
-    #currently this is only used once so it should potentially be merged into the event command
     async def which_raid_question(self, user):
         #declare global variable used in function
         global mycursor
@@ -225,7 +225,124 @@ class helper_cogs(commands.Cog, name='Utilities'):
         mycursor.execute(sql, val)
         mydb.commit()
 
-    
+    # helper utility to change the raid time.
+    async def change_raid_time(self, user, raid_id, new_time):
+        global mycursor
+
+        #create DM channel for user, creating now so except clauses can also use.
+        await user.create_dm()
+
+        try:
+            #if the input is invalid it will throw either ParserError, ValueError, or Overflow Error
+            raid_time = parse(new_time, fuzzy=True)
+
+            #update DB with "when" value
+            sql = "UPDATE raid_plan SET time = %s WHERE idRaids = %s"
+            val = (f'{raid_time.strftime("%I:%M %p %m/%d")}', raid_id)
+            mycursor.execute(sql, val)
+
+            #DM user that raid setup is complete
+            await user.dm_channel.send(f'raid {raid_id} rescheduled to {raid_time.strftime("%I:%M %p %m/%d")}')
+
+            #edit raid post to show new data
+            await self.print_raid(raid_id)
+
+        #catching the error handling to notify user if their input was invalid
+        except ParserError:
+            await user.dm_channel.send(f'not a date time input, please try again')
+        except ValueError:
+            await user.dm_channel.send(f'invalid input, please try again')
+        except OverflowError:
+            await user.dm_channel.send(f'date time values exceed possible values, please try again')
+
+    # helper utility to create raid
+    async def create_raid(self, raid_number: int, raid_time: str):
+        global mycursor
+        global mydb
+        global raid_chan_code
+
+        #create raid post
+        raid_chan = self.bot.get_channel(raid_chan_code)
+        response = f'@here let\'s raid!'
+        message = await raid_chan.send(response)
+
+        #insert raid into DB, currently only setting Raid key and message ID
+        sql = "INSERT INTO raid_plan (`message_id`, `what`, `time`) VALUES (%s, %s, %s)"
+        val = (message.id, raid_number, f'{raid_time.strftime("%I:%M %p %m/%d")}')
+        mycursor.execute(sql, val)
+        mydb.commit()
+
+        raid_setup_id = mycursor.lastrowid
+
+        await self.print_raid(raid_setup_id)
+
+    # helper utility to query the DB
+    async def query_db(self, query: str):
+        global mycursor
+
+        # query DB and grab results
+        mycursor.execute(f'SELECT COUNT(*) FROM raid_info')
+        sqlreturn = mycursor.fetchall()
+
+        # return results
+        return sqlreturn
+
+    # helper utility to create Raid notification posts
+    async def raid_notifiation_check(self):
+        global mycursor
+        global raid_chan_code
+
+        #grab current time.
+        now = datetime.now()
+
+        #pull current information on raids and times.
+        mycursor.execute(f'SELECT idRaids, time, prime_one, prime_two, prime_three, prime_four, prime_five, prime_six, back_one, back_two, notify_message_ID FROM raid_plan WHERE idRaids IS NOT Null')
+        sqlreturn = mycursor.fetchall()
+
+        for i in range(len(sqlreturn)):
+            if(sqlreturn[i][1]is not None):
+                #converting time to a dateutil object to allow comparison
+                raid_time = parse(sqlreturn[i][1], fuzzy=True) 
+
+                #raid_id will be used repeatedly so setting it to a variable
+                raid_id = sqlreturn[i][0]
+
+                #check if raid is starting under 70 minutes from now and does not have a notification message already
+                if (raid_time <= (now + timedelta(minutes = 70))) and sqlreturn[i][10] is None:
+        
+                    #creating int value so the function knows how many people are in the raid
+                    raid_members = 0
+
+                    #beginning of notification message
+                    notify = f'Notification: Raid {raid_id} is starting soon. If you are tagged then you are currently scheduled to raid.\n'
+                    
+                    #adding users to notification message and checking how many people we have
+                    for ii in range(8):
+                        if(sqlreturn[i][ii+2] != None and raid_members < 6):
+                            raid_members += 1
+                            notify =  notify + f'<@{sqlreturn[i][ii+2]}> '
+
+                    #adding a @here mention if we are missing people
+                    if (raid_members < 6):
+                        notify = notify + f'\n@here we still need {6-raid_members} fireteam member(s) for the raid.'
+                    
+                    #notify everyone in the raid and ping @here if we need someone
+                    raid_chan = self.bot.get_channel(raid_chan_code)
+                    message = await raid_chan.send(notify)
+
+                    #get raid post message object and set global variable
+                    notify_message = message
+
+                    #add notify message ID to DB
+                    sql = "UPDATE raid_plan SET notify_message_ID = %s WHERE idRaids = %s"
+                    val = (notify_message.id,  raid_id)
+                    mycursor.execute(sql, val)
+                    mydb.commit()
+
+                #check to see if raid started over 30 minutes ago, if so, delete
+                elif (raid_time + timedelta(minutes = 30) < now):
+                    await self.delete_raid(raid_id)
+
 
 def setup(bot):
     bot.add_cog(helper_cogs(bot))
